+++ src/onegov.ballot/onegov/ballot/tests/models/test_election.py
@@ -765,9 +765,9 @@ def test_election_clear_results(session):
     assert election.candidates.all() == []
     assert election.results.all() == []
 
-    assert session.query(Candidate).count() == 0
-    assert session.query(CandidateResult).count() == 0
-    assert session.query(ElectionResult).count() == 0
+    assert session.query(Candidate).first() is None
+    assert session.query(CandidateResult).first() is None
+    assert session.query(ElectionResult).first() is None
 
 
 def test_election_has_results(session):